Implement FW CAT command
The FW CAT command allows the computer to query the filter settings of
the radio, is currently being queried by WSJT-X when operating in WSPR
mode and makes the radio crash after a couple of hours because two error
response are returned in the same buffer. With this command implemented
the radio does not crash anymore.